requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
pymupdf==1.25.3
httpx==0.28.1
numpy==2.2.2
python-dotenv==1.0.1
